moving query executor
