+++ ee/hogai/README.md
@@ -133,17 +133,17 @@ NOTE: this won't extend query types generation. For that, talk to the PostHog AI
 
 ### Adding a new query type
 
-1. **Update the query executor and formatters** (`@ee/hogai/graph/query_executor/`):
-   - Add a new formatter class in `query_executor/format/` that implements query result formatting for AI consumption. Make sure it's imported and exported from `query_executor/format/__init__.py`. See below (Step 3) for more information.
-   - Add formatting logic to `_compress_results()` method in `query_executor/query_executor.py`:
+1. **Update the query executor and formatters** (`@ee/hogai/context/insight/`):
+   - Add a new formatter class in `context/insight/format/` that implements query result formatting for AI consumption. Make sure it's imported and exported from `context/insight/format/__init__.py`. See below (Step 3) for more information.
+   - Add formatting logic to `_compress_results()` method in `context/insight/query_executor.py`:
 
      ```python
      elif isinstance(query, YourNewAssistantQuery | YourNewQuery):
          return YourNewResultsFormatter(query, response["results"]).format()
      ```
 
-   - Add example prompts for your query type in `query_executor/prompts.py`, this explains to the LLM the query results formatting
-   - Update `_get_example_prompt()` method in `query_executor/nodes.py` to handle your new query type:
+   - Add example prompts for your query type in `context/insight/prompts.py`, this explains to the LLM the query results formatting
+   - Update `get_example_prompt()` function in `context/insight/query_executor.py` to handle your new query type:
 
      ```python
      if isinstance(viz_message.answer, YourNewAssistantQuery):
