+++ ee/hogai/chat_agent/query_executor/test/test_nodes.py
@@ -30,7 +30,7 @@ from posthog.api.services.query import process_query_dict
 
 from ee.hogai.artifacts.manager import ArtifactManager
 from ee.hogai.chat_agent.query_executor.nodes import QueryExecutorNode
-from ee.hogai.chat_agent.query_executor.prompts import (
+from ee.hogai.context.insight.prompts import (
     FUNNEL_STEPS_EXAMPLE_PROMPT,
     FUNNEL_TIME_TO_CONVERT_EXAMPLE_PROMPT,
     FUNNEL_TRENDS_EXAMPLE_PROMPT,
@@ -51,7 +51,7 @@ class TestQueryExecutorNode(ClickhouseTestMixin, NonAtomicBaseTest):
         self.conversation = Conversation.objects.create(user=self.user, team=self.team)
         self.artifact_manager = ArtifactManager(self.team, self.user)
 
-    @patch("ee.hogai.chat_agent.query_executor.query_executor.process_query_dict", side_effect=process_query_dict)
+    @patch("ee.hogai.context.insight.query_executor.process_query_dict", side_effect=process_query_dict)
     async def test_node_legacy_viz_message_runs(self, mock_process_query_dict):
         node = QueryExecutorNode(self.team, self.user)
         new_state = await node.arun(
@@ -97,7 +97,7 @@ class TestQueryExecutorNode(ClickhouseTestMixin, NonAtomicBaseTest):
         self.assertFalse(new_state.root_tool_insight_plan)
         self.assertFalse(new_state.root_tool_insight_type)
 
-    @patch("ee.hogai.chat_agent.query_executor.query_executor.process_query_dict", side_effect=process_query_dict)
+    @patch("ee.hogai.context.insight.query_executor.process_query_dict", side_effect=process_query_dict)
     async def test_node_runs(self, mock_process_query_dict):
         node = QueryExecutorNode(self.team, self.user)
 
@@ -154,7 +154,7 @@ class TestQueryExecutorNode(ClickhouseTestMixin, NonAtomicBaseTest):
         self.assertFalse(new_state.root_tool_insight_type)
 
     @patch(
-        "ee.hogai.chat_agent.query_executor.query_executor.process_query_dict",
+        "ee.hogai.context.insight.query_executor.process_query_dict",
         side_effect=ValueError("You have not glibbled the glorp before running this."),
     )
     async def test_node_handles_internal_error(self, mock_process_query_dict):
@@ -196,7 +196,7 @@ class TestQueryExecutorNode(ClickhouseTestMixin, NonAtomicBaseTest):
         self.assertIsNotNone(msg.id)
 
     @patch(
-        "ee.hogai.chat_agent.query_executor.query_executor.process_query_dict",
+        "ee.hogai.context.insight.query_executor.process_query_dict",
         side_effect=ValidationError(
             "This query exceeds the capabilities of our picolator. Try de-brolling its flim-flam."
         ),
@@ -264,7 +264,7 @@ class TestQueryExecutorNode(ClickhouseTestMixin, NonAtomicBaseTest):
 
     async def test_fallback_to_json(self):
         node = QueryExecutorNode(self.team, self.user)
-        with patch("ee.hogai.chat_agent.query_executor.query_executor.process_query_dict") as mock_process_query_dict:
+        with patch("ee.hogai.context.insight.query_executor.process_query_dict") as mock_process_query_dict:
             mock_process_query_dict.return_value = QueryStatus(
                 id="test", team_id=self.team.pk, query_async=True, complete=True, results=[{"test": "test"}]
             )
@@ -308,7 +308,7 @@ class TestQueryExecutorNode(ClickhouseTestMixin, NonAtomicBaseTest):
             self.assertIsNotNone(msg.id)
 
     def test_get_example_prompt(self):
-        from ee.hogai.chat_agent.query_executor.query_executor import get_example_prompt
+        from ee.hogai.context.insight.query_executor import get_example_prompt
 
         # Test Trends Query
         trends_query = AssistantTrendsQuery(series=[AssistantTrendsEventsNode()])
