+++ ee/hogai/context/insight/format/funnel.py
