+++ ee/hogai/context/insight/format/retention.py
