+++ ee/hogai/context/insight/format/revenue_analytics.py
