+++ ee/hogai/context/insight/format/test/test_funnel.py
