+++ ee/hogai/context/insight/format/test/test_revenue_analytics.py
