+++ ee/hogai/context/insight/format/test/test_trends.py
