+++ ee/hogai/context/insight/format/test/test_utils.py
