+++ ee/hogai/context/insight/format/trends.py
