+++ ee/hogai/context/insight/format/utils.py
