+++ ee/hogai/context/insight/prompts.py
